graphqlgen implemented for type generation
Issues: not sufficient for usage with typeorm.
Graphqlgen is designed for use with Prisma as the ORM.
It ONLY allows for "Models" that are interfaces (where "Models" are the parents passed down to resolvers), meaning we cannot return Entities from parent resolvers. The workarounds are messy. There ought to be another codegen tool that doesn't require models to be interfaces